+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,35 +2,27 @@ package roomescape.controller;
 
 import jakarta.validation.Valid;
 import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.Login;
-import roomescape.controller.dto.PaymentApproveRequest;
 import roomescape.controller.dto.UserReservationSaveRequest;
 import roomescape.controller.dto.UserReservationViewResponse;
 import roomescape.controller.dto.UserReservationViewResponses;
 import roomescape.service.ReservationService;
 import roomescape.service.dto.request.LoginMember;
-import roomescape.service.dto.request.ReservationSaveRequest;
 import roomescape.service.dto.response.ReservationResponse;
-
-import java.util.List;
 
 @RestController
 public class ReservationController {
 
     private final ReservationService reservationService;
-    private final PaymentController paymentController;
 
-    public ReservationController(
-            ReservationService reservationService,
-            PaymentController paymentController
-    ) {
+    public ReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
-        this.paymentController = paymentController;
     }
 
     @PostMapping("/reservations")
@@ -38,11 +30,7 @@ public class ReservationController {
             @Login LoginMember member,
             @RequestBody @Valid UserReservationSaveRequest userReservationSaveRequest
     ) {
-        PaymentApproveRequest paymentApproveRequest = PaymentApproveRequest.from(userReservationSaveRequest);
-        paymentController.approve(new HeaderGeneratorImpl(), paymentApproveRequest);
-
-        ReservationSaveRequest reservationSaveRequest = userReservationSaveRequest.toReservationSaveRequest(member.id());
-        ReservationResponse reservationResponse = reservationService.saveReservation(reservationSaveRequest);
+        ReservationResponse reservationResponse = reservationService.saveUserReservation(member, userReservationSaveRequest);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
                 .body(reservationResponse);
     }
