+++ src/main/java/roomescape/service/PaymentService.java
@@ -1,28 +1,29 @@
-package roomescape.controller;
+package roomescape.service;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
+import roomescape.controller.HeaderGenerator;
+import roomescape.controller.PaymentApproveResponse;
 import roomescape.controller.dto.PaymentApproveRequest;
 
-@RestController
-public class PaymentController {
+@Service
+public class PaymentService {
     private static final String PAYMENT_APPROVE_ENDPOINT = "https://api.tosspayments.com/v1/payments/confirm";
 
     @Value("${payment.secret-key}")
     private String secretKey;
     private final RestTemplate restTemplate;
 
-    public PaymentController(RestTemplate restTemplate) {
+    public PaymentService(RestTemplate restTemplate) {
         this.restTemplate = restTemplate;
     }
 
-    public ResponseEntity<PaymentApproveResponse> approve(HeaderGenerator headerGenerator, PaymentApproveRequest paymentApproveRequest) {
+    public PaymentApproveResponse pay(HeaderGenerator headerGenerator, PaymentApproveRequest paymentApproveRequest) {
         HttpHeaders headers = headerGenerator.generate();
         headers.setBasicAuth(encodeSecretKey());
         HttpEntity httpEntity = new HttpEntity(paymentApproveRequest, headers);
@@ -31,7 +32,7 @@ public class PaymentController {
                 PAYMENT_APPROVE_ENDPOINT,
                 httpEntity,
                 PaymentApproveResponse.class
-        );
+        ).getBody();
     }
 
     private String encodeSecretKey() {
