+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -242,7 +242,7 @@ class ReservationControllerTest extends IntegrationTestSupport {
                 "themeId", 1L
         );
 
-        Mockito.when(paymentController.approve(any(), any()))
+        Mockito.when(paymentService.pay(any(), any()))
                 .thenThrow(RoomEscapeBusinessException.class);
 
         RestAssured.given().log().all()
